requirements.txt
flask
pandas
scikit-learn
openai
python-dotenv
requests
joblib